Put actions and mutations and state in their own files Use computed property name for actions and mutations
